// Backend/src/interfaces/user.interfaces.ts
export interface SocialMediaProfileLinks {
     platform: string;
     link: string;
 }
 
 export interface FollowersCount {
     platform: string;
     count: number;
 }
 
 export interface PreviousSponsorships {
      company_name: string;
      
 }
 
 export interface ContentType {
     content: string;
 }
 
 